requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
motor==3.3.2
pymongo==4.6.0
celery==5.3.4
redis==5.0.1
boto3==1.34.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
